Refactor Duke to MeetingJio
