--- src/main/java/seedu/duke/events/Lesson.java
@@ -1,16 +1,0 @@
-package seedu.duke.events;
-
-public class Lesson extends Event {
-    public Lesson(String name, String title, String day, int startTime, int endTime, String mode) {
-        super(name, title, day, startTime, endTime, mode);
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (!(obj instanceof Lesson)) {
-            return false;
-        }
-        Lesson lesson = (Lesson) obj;
-        return super.equals(lesson);
-    }
-}
