--- src/main/java/seedu/duke/exceptions/DuplicateEventException.java
@@ -1,4 +1,0 @@
-package seedu.duke.exceptions;
-
-public class DuplicateEventException extends Exception {
-}
