--- src/main/java/seedu/duke/exceptions/InvalidDayException.java
@@ -1,4 +1,0 @@
-package seedu.duke.exceptions;
-
-public class InvalidDayException extends Exception {
-}
