--- src/main/java/seedu/duke/exceptions/InvalidModeException.java
@@ -1,4 +1,0 @@
-package seedu.duke.exceptions;
-
-public class InvalidModeException extends Exception {
-}
