--- src/main/java/seedu/duke/exceptions/InvalidTimeException.java
@@ -1,4 +1,0 @@
-package seedu.duke.exceptions;
-
-public class InvalidTimeException extends Exception {
-}
