--- src/main/java/seedu/duke/exceptions/MissingValueException.java
@@ -1,5 +1,0 @@
-package seedu.duke.exceptions;
-
-public class MissingValueException extends Exception {
-
-}
