--- src/main/java/seedu/duke/exceptions/OverlappingEventException.java
@@ -1,4 +1,0 @@
-package seedu.duke.exceptions;
-
-public class OverlappingEventException extends Exception {
-}
