+++ src/main/java/seedu/meetingjio/exceptions/DuplicateEventException.java
@@ -1,0 +1,4 @@
+package seedu.meetingjio.exceptions;
+
+public class DuplicateEventException extends Exception {
+}
