+++ src/main/java/seedu/meetingjio/exceptions/InvalidDayException.java
@@ -1,0 +1,4 @@
+package seedu.meetingjio.exceptions;
+
+public class InvalidDayException extends Exception {
+}
