+++ src/main/java/seedu/meetingjio/exceptions/InvalidModeException.java
@@ -1,0 +1,4 @@
+package seedu.meetingjio.exceptions;
+
+public class InvalidModeException extends Exception {
+}
