+++ src/main/java/seedu/meetingjio/exceptions/InvalidTimeException.java
@@ -1,0 +1,4 @@
+package seedu.meetingjio.exceptions;
+
+public class InvalidTimeException extends Exception {
+}
