+++ src/main/java/seedu/meetingjio/exceptions/MissingValueException.java
@@ -1,0 +1,5 @@
+package seedu.meetingjio.exceptions;
+
+public class MissingValueException extends Exception {
+
+}
