+++ src/main/java/seedu/meetingjio/exceptions/OverlappingEventException.java
@@ -1,0 +1,4 @@
+package seedu.meetingjio.exceptions;
+
+public class OverlappingEventException extends Exception {
+}
