+++ src/test/java/seedu/meetingjio/MeetingJioTest.java
@@ -1,0 +1,12 @@
+package seedu.meetingjio;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import org.junit.jupiter.api.Test;
+
+class MeetingJioTest {
+    @Test
+    public void sampleTest() {
+        assertTrue(true);
+    }
+}
